tidy up IGRINS-2 recipes
