+++ geminidr/igrins/recipes/sq/recipes_STANDARD_SPECT.py
@@ -1,0 +1,10 @@
+"""
+Recipes available to data with tags ['IGRINS-2', 'SPECT', 'XD'', 'STANDARD].
+Default is "reduceTelluric".
+"""
+from .recipes_SPECT import reduceTelluric
+
+
+recipe_tags = {'IGRINS-2', 'SPECT', 'XD', 'STANDARD'}
+
+_default = reduceTelluric
